feat: Replace SVG browser logos with PNG for better performance and compatibility
